requirements.txt
openai-whisper~=20240930
ffmpeg
torch
python-dotenv
google-genai